git: filter out .git from artifact tgz
The .git/object/* files are all read-only, so when they are re-deployed
to an executor it fails to overwrite them. One solution is to unlink
them before overwriting them, but I think the better solution is just
to filter the entire repo out of the artifact since it solves the
problem and also potentially greatly reduces disk usage and bandwidth.