executor: fix handling of generic requests
The callback given to a strong-control-channel request is only given a
single argument: the response. Since the error, if any, is given as a
property of that response we need to extract it to behave like a normal
node-back. Basically this wrapper wasn't doing what it was supposed to
be doing, but now it is.

Bandwidth usage aside, this seems to be the main thing that was
preventing heap snapshots and cpu profiling from working.